requirements.txt
streamlit==1.34.0
langchain==0.1.20
langgraph==0.0.37
langchain-openai==0.1.6
python-dotenv==1.0.1
openai==1.25.0
PyPDF2==3.0.1
tiktoken==0.6.0
faiss-cpu==1.7.4
pandas==2.2.2
scikit-learn==1.4.2
